src/requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.10
Django==5.1.6
django-crispy-forms==2.3
django-filter==24.3
pillow==11.1.0
python-stdnum==1.20
sqlparse==0.5.3
tzdata==2025.1
